+++ lib/src/models/events.dart
@@ -18,5 +18,6 @@ enum MapBoxEvent {
   banner_instruction,
   on_arrival,
   failed_to_reroute,
-  reroute_along
+  reroute_along,
+  annotation_tapped
 }